Made changes suggested by clang-tidy

+++ Utility.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#include <random>
 #include <iostream>
+#include <random>
 #include <vector>
 
 template <class T> 
@@ -94,24 +94,24 @@ double tolerancia_fisica(double numsigma)
 {
 	return exp(-numsigma*numsigma/2.0);
 }*/
-const unsigned long long KB = 1024;
-const unsigned long long MB = KB*1024;
-const unsigned long long GB = MB*1024;
+const uint64_t KB = 1024;
+const uint64_t MB = KB*1024;
+const uint64_t GB = MB*1024;
 
 
 #if defined(__unix__) || defined(__unix) || (defined(__APPLE__) && defined(__MACH__))
 	#include <unistd.h>
 
-	inline unsigned long long getTotalSystemMemory()
+	inline uint64_t getTotalSystemMemory()
 	{
-		unsigned long long  pages = sysconf(_SC_PHYS_PAGES);
-		unsigned long long  page_size = sysconf(_SC_PAGE_SIZE);
+		uint64_t  pages = sysconf(_SC_PHYS_PAGES);
+		uint64_t  page_size = sysconf(_SC_PAGE_SIZE);
 		return pages * page_size;
 	}
 #else 
 	#include <windows.h>
 
-	inline unsigned long long getTotalSystemMemory()
+	inline uint64_t getTotalSystemMemory()
 	{
 		MEMORYSTATUSEX status;
 		status.dwLength = sizeof(status);
